+++ app/server/lib/steps/satellite.js
