+++ package.json
@@ -2,35 +2,33 @@
   "name": "course-terrain-tool",
   "version": "1.0.0",
   "description": "An open source tool to find and process terrain data for use in the OPCD golf course design process",
+  "author": {
+    "name": "dudewheresmycode",
+    "url": "https://github.com/dudewheresmycode"
+  },
   "workspaces": [
     "client"
   ],
   "type": "module",
   "main": "app/main.js",
   "build": {
-    "asar": true,
     "productName": "Course Terrain Tool",
     "icon": "./resources/icon.icns",
     "appId": "com.opengolfsim.ctt",
     "mac": {
       "category": "public.app-category.developer-tools"
-    },
-    "extraResources": [
-      "./server",
-      "./client/dist",
-      "./resources"
-    ]
+    }
   },
   "scripts": {
-    "start": "CTT_ELECTRON=1 electron .",
+    "start": "electron .",
     "app:dir": "electron-builder --dir",
     "app:dist": "electron-builder",
     "test": "echo \"Error: no test specified\" && exit 1",
-    "client": "CLIENT_DEV_MODE=1 concurrently --kill-others  'npm run debug --workspace=client' 'npm run -w app start'",
+    "client": "CLIENT_DEV_MODE=1 concurrently --kill-others  'npm run debug --workspace=client' 'npm run start'",
     "_server": "npm start --workspace=server",
     "build": "npm run build --workspace=client",
     "_develop": "CTC_DEBUG=1 concurrently --kill-others 'npm run debug --workspace=client' 'npm start --workspace=server'",
-    "_start_server": "npm start --workspace=server",
+    "standalone": "node app/standalone.js",
     "_app_deps": "electron-builder install-app-deps"
   },
   "keywords": [],
