--- server/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "ctt-server",
-  "type": "module",
-  "main": "index.js",
-  "scripts": {
-    "start": "node index.js",
-    "test": "node ./test/job.js"
-  },
-  "dependencies": {
-    "axios": "^1.7.9",
-    "express": "^4.21.2",
-    "express-ws": "^5.0.2",
-    "p-map": "^7.0.3",
-    "sharp": "^0.33.5",
-    "staticmaps": "^1.13.1",
-    "dotenv": "^16.4.7"
-  }
-}
